requirements.txt
python_telegram_bot
matplotlib
sqlalchemy